RPC Binder: limit experimental wire protocol use
Only allow experimental wire protocol when:
- we're on Android
- it's not a release build

This is intended to avoid things in the wild accidentally
freezing the experimental wire protocol. If such a thing
happened, we'd have to stop using the existing experimental
wire protocol version and increment it. Worse though, it
may de facto freeze the experimental version at that time,
or even cause RPC binder to fail in certain cases.

One downside of this is that when we make changes to the
RPC binder wire protocol, that we would no longer be
able to test it. In order to avoid that problem, we could
add another way to enable it at that time. However, to
play things on the safe side, disabling by default
for now.

Fixes: 305983144
Test: binderRpcTest
Change-Id: Ieb23cc11d6f3c5b77545dcfd7621f1464dd98f0f