Do not rely on android_is_in_vendor_process
android_is_in_vendor_process is used to check if the function is called
from the system process. However, implementation is incorrect because it
is called only when VNDK is not enabled, while
android_is_in_vendor_process only works when VNDK is enabled. Instead,
use __ANDROID_VENDOR__ and __ANDROID_PRODUCT__ flag to check if it is
compiled as vendor / product variant.

Bug: 300366609
Test: aosp cf build succeded
Change-Id: I7cb904fe9910e5325132c68ca584eb964c84a79b